+++ src/rex.deploy/NEWS.rst
@@ -10,6 +10,7 @@
 
 * Support for calculated fields.
 * Identity-to-text conversion.
+* Added backlink introspection.
 
 
 2.5.0 (2016-02-29)